WIP: wishlist feature
