+++ src/components/Wishlist/StyledComponents.js
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import { List } from "antd";
+
+export const DataList = styled(List)`
+	&& {
+		margin: 0 10px 70px 10px;
+	}
+`;
